Thu, Jan 11, 2024, 7:01 PM +01:00

+++ CS50/Week 4/filter-more/helpers.c
@@ -1,5 +1,5 @@
 #include "helpers.h"
-#include <math.h> 
+#include <math.h>
 #include <stdio.h> // BORRAR DESPUÉS
 // Convert image to grayscale
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
@@ -18,7 +18,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
                 printf("numero: %u\n", image[i][j].rgbtBlue);
                 printf("numero: %u\n", image[i][j].rgbtGreen);
                 printf("numero: %u\n", image[i][j].rgbtRed);
-                printf("numero: %f\n", round((float)suma/3));
+                printf("numero: %u\n", (BYTE)round((float)suma/3));
             }
         }
     }
